requirements.txt
discord.py
qrcode
pillow
python-dotenv